+++ mllib/src/main/scala/org/apache/spark/ml/evaluation/RankingEvaluator.scala
@@ -26,6 +26,7 @@ import org.apache.spark.mllib.evaluation.RankingMetrics
 import org.apache.spark.sql.Dataset
 import org.apache.spark.sql.functions._
 import org.apache.spark.sql.types._
+import org.apache.spark.sql.util.SchemaUtils
 
 /**
  * :: Experimental ::
