+++ mllib/src/main/scala/org/apache/spark/ml/feature/BucketedRandomProjectionLSH.scala
@@ -18,9 +18,7 @@
 package org.apache.spark.ml.feature
 
 import scala.util.Random
-
 import org.apache.hadoop.fs.Path
-
 import org.apache.spark.annotation.Since
 import org.apache.spark.ml.linalg._
 import org.apache.spark.ml.param._
@@ -29,6 +27,7 @@ import org.apache.spark.ml.util._
 import org.apache.spark.mllib.util.MLUtils
 import org.apache.spark.sql.Row
 import org.apache.spark.sql.types.StructType
+import org.apache.spark.sql.util.SchemaUtils
 
 /**
  * Params for [[BucketedRandomProjectionLSH]].
