+++ mllib/src/main/scala/org/apache/spark/ml/feature/Normalizer.scala
@@ -73,7 +73,7 @@ class Normalizer @Since("1.4.0") (@Since("1.4.0") override val uid: String)
     if ($(inputCol).nonEmpty && $(outputCol).nonEmpty) {
       val size = AttributeGroup.fromStructField(schema($(inputCol))).size
       if (size >= 0) {
-        outputSchema = SchemaUtils.updateAttributeGroupSize(outputSchema,
+        outputSchema = AttributeGroup.updateAttributeGroupSize(outputSchema,
           $(outputCol), size)
       }
     }
