+++ mllib/src/main/scala/org/apache/spark/ml/feature/QuantileDiscretizer.scala
@@ -26,6 +26,7 @@ import org.apache.spark.ml.param.shared._
 import org.apache.spark.ml.util._
 import org.apache.spark.sql.Dataset
 import org.apache.spark.sql.types.StructType
+import org.apache.spark.sql.util.SchemaUtils
 
 /**
  * Params for [[QuantileDiscretizer]].
