+++ mllib/src/main/scala/org/apache/spark/ml/stat/ChiSquareTest.scala
@@ -19,7 +19,7 @@ package org.apache.spark.ml.stat
 
 import org.apache.spark.annotation.Since
 import org.apache.spark.ml.linalg.{Vector, Vectors, VectorUDT}
-import org.apache.spark.ml.util.SchemaUtils
+import org.apache.spark.sql.util.SchemaUtils
 import org.apache.spark.mllib.linalg.{Vectors => OldVectors}
 import org.apache.spark.mllib.stat.test.{ChiSqTest => OldChiSqTest}
 import org.apache.spark.sql.{DataFrame, Row}
